lab_4 copy/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "httpx>=0.28",
    "nltk>=3.9.1",
    "greenlet>=3.3.2",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "ruff>=0.15.6",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.48",
    "striprtf>=0.0.29",
    "uvicorn>=0.41.0",
    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
